python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aimo-network"
version = "0.1.0"
description = "Python SDK for AiMo Network - a decentralized AI inference marketplace"
readme = "README.md"
license = "ISC"
requires-python = ">=3.10"
authors = [{ name = "AiMo Network" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = ["httpx>=0.27"]

[project.optional-dependencies]
solana = ["solders>=0.21", "base58>=2.1", "x402[svm]>=2.2"]
evm = ["eth-account>=0.13", "x402[evm]>=2.2"]
openai = ["openai>=1.50"]
x402 = ["x402[httpx]>=2.2"]
all = ["aimo-network[solana,evm,openai,x402]"]
dev = [
    "aimo-network[all]",
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "python-dotenv>=1",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.urls]
Homepage = "https://aimo.network"
Repository = "https://github.com/aimonetwork/aimo-sdk"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
